Some cosmetic code movements and comments
